Create AnnotationItems immediately when parsing types
Previously, the annotations would initially be collated as `String`s
containing the source representation of the annotation and then
converted into an actual `AnnotationItem` while being added to the
`DefaultTypeModifiers`. However, during that period when they were
strings it was necessary to check to see if they were a nullability
annotation.

This change creates them as `AnnotationItem`s straight away avoiding
the need to manage them as strings and taking advantage of the
`AnnotationItem` methods to check for nullability annotations which
cache the result in an `AnnotationInfo` for later use.

This should not negatively impact performance as it does not change
which `AnnotationItem`s are created, only when.

Bug: 325691395
Test: ./gradlew
Change-Id: Ia1b305b85b8f79cb5cfcae50d99ae2c71c750950