Add PsiFileLocation to wrap a PsiElement
Previously, the `PsiItem.fileLocation` (and the `PsiConstructorItem`
override) were abstract properties that created a `FileLocation` on
demand as the computation of path and line number was not trival and
was rarely needed. This change moves the computation into a new
`FileLocation` subclass `PsiFileLocation` that wraps the `PsiElement`
from which the path and line number could be computed.

Bug: 325962341
Test: ./gradlew
Change-Id: I01e21883aa7f3211e9010a070b087907ba7f0bf5