Pull isAssignableFromWithoutUnboxing() up into TypeItem
Implements the `isAssignableFromWithoutUnboxing()` on `TypeItem` using
the "fallback to a limited text based check" logic and keeps the
`PsiTypeItem` override. That allows the `PsiTypeItem` to be marked as
`internal`.

Follow-up changes will add test suite tests for them (based off the
existing tests), implement it consistently for the models and remove
the Psi specific implementation.

Bug: 324196754
Test: ./gradlew
Change-Id: Iaf09010079d0b5dc03e4d625b2f0a3f95c8da01b