Clear final on enum methods in a final class
Previously, it would clear the `final` modifier on methods that are
not "in an enum" and not "an enum synthetic method" but it should have
cleared it on methods that are not "in an enum and are synthetic". As
the `isEnumSyntheticMethod()` already checks to make sure it is in an
enum class that can be simplified to just `!isEnumSyntheticMethod()`.
This change corrects that and updates the affected tests.

This will have a minor impact on API signature files downstream.

Bug: 326310840
Test: m checkapi
Change-Id: Ie00e092562f2d9960f579e17300a955fb9a5d857