Create TextMethodItem and TextConstructorItem after parsing
Previously, they were created in the middle of parsing the method and
constructor lines respectively. That made it difficult to optionally
ignore an enum synthetic methods. So, this change moves the creation
of `TextMethodItem` after the end of line `;` token has been parsed.
It does the same for `TextConstructorItem` to be consistent.

Bug: 326310840
Test: ./gradlew
Change-Id: I19e0cdc405e5a06e9dcf914e7f356612f3348e28