Implement Item.toString() consistently across models
As it is not possible for an interface to implement methods on `Any`
like `toString()` this adds an `Item.toStringForItem()` method and then
implements `DefaultItem.toString()` to call it. That allows the
subinterfaces of `Item` to provide a default implementation of
`toStringForItem()` that is consistent across all models.

The previous implementations were inconsistent, e.g. text properties
used "Field" as the prefix, psi properties used "field" as the prefix
when they should both have used "property". Fixing them caused problems
with the AndroidX build as the prefix is part of the element id used to
find matching entries in the baseline. Changing the prefix meant that
existing issues in old baseline files were not being found. This
temporarily compensates for that in the `Baseline` code until the
baseline files can be updated.

Bug: 325454211
Test: ./gradlew
Change-Id: Id8a08168d2d09a805c58988c57bc65b82ecd6aaa