Only check references if an element is not effectivelyDeprecated
This previously checked if the method or containing class was not
originallyDeprecated. That isn't the case for inner classes that
inherit deprecated status from an outer class.

That worked because this check currently does not run on most inner
classes. The followup change fixes that, so this change is needed to
prevent incorrect ReferencesDeprecated errors.

Bug: 326571399
Bug: 315206833
Test: Added test that the warnings don't come up for an inner class.
This would currently pass regardless, but without this change, the
followup change would make it fail.

Change-Id: I8f47cc37cb493800a8ac9411df851b9a5c8823d7