Add some simple tests for a data class
Data classes appear to have some special handling in the IntelliJ
Kotlin -> Psi mapping which causes some discrepancies between the type
nullability and the nullability annotations added to the parameters of
the `copy` method. This change adds some tests to allow that handling
to be tested and the discrepancies investigated and worked around.

Bug: 325454211
Test: ./gradlew
Change-Id: I61972e9f6c00f4e0efe551a95e608b4f00bf7271