Use firstOrNull() in List<AnnotationItem>.typeNullability
Previously, it used `singleOrNull()` but that does not behave as
required. The intent was that it would fail if there was more than one,
return the single item if there was one and otherwise return `null`.
Unfortunately, it actually returns `null` if there is not exactly one
item. That means that a list containing multiple nullability
annotations would behave as if it contained none which is very
confusing.

This change switches to use `firstOrNull()` as that will be slightly
faster (as it does not have to iterate over the whole, albeit generally
a short, list) but mainly because the behavior makes more sense.

Bug: 325454211
Test: ./gradlew
Change-Id: I7aa995503539bf7c58cc054426b6edfa7e6e55de