Add CustomizableParameterizedRunner
The JUnit 4 `Parameterized` test runner is really useful but is limited
in how it can be customized. This adds a new
`CustomizableParameterizedRunner` class that will be the basis for some
improvements to the model test running that will simplify the writing
of tests parameterized by `ModelSuiteRunner`, particularly those that
are also parameterized in other ways. This is needed to begin migrating
main metalava tests to run against multiple source model providers.

Bug: 326649406
Test: ./gradlew
Change-Id: If6e4e75c7ce34e274aa735c432295753a39b4ce5