Inject ModelProviderTestInfo into ModelProvideTestAware directly
Previously, the `ModelProviderTestInfo` would be in the argument list
in `TestParameters` and `BlockJUnit4ClassRunnerWithParameters` would
then inject that into a field annotated with `@Parameter(0)`. The
problem with that is when migrating an existing test that is already
`Parameterized`, it would require the existing `@Parameter(n)`
annotations to be switched to `@Parameter(n+1)`, as otherwise there
would be multiple `@Parameter(0)` annotated fields which would cause a
test failure.

By extracting the `ModelProviderTestInfo` and injecting it directly it
avoids having to modify the existing `@Parameter` annotations. That
makes switching over to run tests against multiple model providers
easier.

Bug: 326649406
Test: ./gradlew
Change-Id: I5f9690426a2248c93093060ecabe5a68783fc60f