Add ModelProviderAwareTest
Part of the work to generalize support for running model provider
tests. Adds an interface for tests to implement and moves and renames
`TestParameters` to be part of that interface.

This moves functionality related to the `ModelProvideTestInfo` from
`BaseModelTest` into a new `ModelProviderAwareTest` that other tests
will be able to extend and use.

Bug: 326649406
Test: ./gradlew
Change-Id: Idcaf484b10e5a953eb202945584bca650b07e50c