Add TestEnvironment to allow tests to pass info to the MainCommand
Previously, in order to pass information from the test to the
`MainCommand` it was necessary to add command line options that were
not documented in the help and so should only be used by tests; options
like `--skip-emit-packages`. However, there was nothing stopping those
options being used from the real command line and it also limited what
could be passed to what could be easily encoded into a command line
option.

This change adds the `TestEnvironment` which can be used to pass any
information through to the `MainCommand`. It uses that to remove the
need for the `--skip-emit-packages` option.

This change revealed a fragility in the tests caused by the
`--api-lint` option. That option will consume the following argument
unless it starts with a `-` or is not a file. It appears that for some
tests the `--skip-emit-packages` option was the only option between
`--api-lint` and the source files. Removing `--skip-emit-packages`
caused those tests to start failing because the `--api-lint` option
was consumed the java source file that followed it and then tried to
parse it as an API signature file which changed.

Usually that is avoided by users of Metalava by making sure that either
`--api-lint` has an API signature file or is immediately following by
another option. However, that is tricky with the tests because the
exact set of arguments passed on the command line is not known. This
change works around that by moving the `--api-lint` options to the end
of the arguments. Bug 327019023 tracks the work to fix `--api-lint`
properly.

Bug: 326649406
Test: ./gradlew
Change-Id: I5dbfa3481fa848ed893a6bc4d742010142ec5062