Allow running test methods from within the IDE
When requested to run a specific method, e.g. `test.pkg.FooTest.foo()`
the IDE will generate something like this:
```
--tests test.pkg.FooTest.foo
```

Any test description that matches will be run which historically has
worked as that matches the default description of the test method.
Unfortunately, change https://r.android.com/2977087 changed the default
description to be `test.pkg.FooTest.foo[psi]` so it no longer matches.
Ideally, the parameter suffix should not prevent the match and so cause
ALL tests for that method to run, in a similar way to how running a
whole class would cause all tests for that class to run. Unfortunately,
that is not how it works and there is no way to modify that behavior.

So, this change tweaks the description generation to avoid adding a
suffix for the default provider, which is currently set to "psi".

Bug: 326649406
Test: ./gradlew
Change-Id: Id37bae08e287bd236eabc5085acb54df88769fda