Switch all subclasses of DriverTest to use DriverTestRunner
Previously, some subclasses of `DriverTest` were already parameterized
and so used `@RunWith(Parameterized::class)` which overrides the
`@RunWith(DriverTestRunner::class)` on `DriverTest`. As a result they
would not set the `DriverTest.sourceModelTestInfo` and so it had to be
initialized. This changes removes the `@RunWith(Parameterized::class)`
on the subclasses and stops initializing `sourceModelTestInfo` turning
it into a `lateinit var`.

There is one slight wrinkle which is that the `@Ignore` annotation
does not work on when specified on a class that is run with
`DriverTestRunner` so this change adds `@RunWith(Parameterized::class)`
to `CheckAndroidApisShard4Test` to work around that issue.

Bug: 326649406
Test: ./gradlew
Change-Id: Iec5be2684a5e58ba16a24d7bce5003825ce2a1bb