Rename Location to IssueLocation
`Location` was a very generic name that did not clearly indicate the
purpose of the class. As the documentation for the class indicates it
contains two forms of "location" information.

It would also be helpful to be able to access the `path` and `line`
"location" information separately to the `baselineKey`. Adding another
form of "location" without clarifying what this one was for would be
extra confusing.

Hence, this change renames `Location` to `IssueLocation`.

Bug: 325962341
Test: ./gradlew
Change-Id: Ib966738650429328966d3753614a2ca78bfc027b