Change Reportable.location() to Reportable.issueLocation val
This change renames `location` to `issueLocation` to match the type and
differentiate it from another "location" to encapsulate the `path` and
`line` properties that will be added in follow-up changes. It also
changes it from a zero-parameter function to an abstract property at
the same time as it will impact the exact same classes and does not add
to much review overhead.

It was switched to `val` as that simplifies access, albeit slightly,
and will reduce overhead for those implementations that want to provide
it as a property while allowing others to still compute it dynamically
as needed.

Bug: 325962341
Test: ./gradlew
Change-Id: I95864ac28b858eff6f31d739f6937205c21a92ae