Annotate all tests that use kotlin(...)
Adds `@RequiresCapabilities(Capability.KOTLIN)` to all tests that use
`kotlin(...)` as there is no point in running a test that is can never
work. That annotation prevents the test being run against Turbine and
as a result 169 tests are removed from the baseline file.

Removes `@FilterByProvider("turbine", EXCLUDE)` from `UastTestK2.kt` as
none of its tests are run on Turbine anyway.

Bug: 326649406
Test: ./gradlew
Change-Id: I9a9acf5504deca917065444c536ab672722ab8f1