Run all instances of a parameterized test method
Previously, the `DriverTestRunner` would try and work around IntelliJ's
inability to run specific methods on a parameterized test that does not
use `@RunWith(org.junit.runners.Parameterized)` by removing the test
suffix containing the arguments, i.e. removing `[....]`. Unfortunately,
that also removes it from the baseline files making it hard to
understand which specific instances is broken and makes it difficult
to run all instances of a specific method.

This change reverts that previous attempt and instead fixes up the
pattern that the IDE sends to Gradle. That works much better as it will
run all the method instances and baseline file clearly indicates which
test instance is failing.

While testing it was found that the pattern IntelliJ uses when running
a specific instance of the `TokenizerTest.check token` method would
match multiple instances. It appears to use wildcards inside the
`[...]`, at the beginning and end, e.g. `[*...*]` and also for some
characters, e.g. `"` gets replaced with `*`. To avoid that this change
adds the index to the test parameters.

Bug: 326649406
Test: ./gradlew
Change-Id: Iab62d9cb93f3a8429d6c23e32fb9ae69c308a2dd