Fix the nullability of a vararg parameter in an inline function
Previously, the `vararg` modifier was not being tracked correctly for
a vararg parameter of an inline reified function as the parameter used
the `LightModifierList` but that was only used for `UMethod`, not
`UParameter`.

This change removes the restriction to only use that for `UMethod` and
updates the test to match.

Bug: 325454211
Test: ./gradlew
Change-Id: Ie52e1376834d5574d2d5a6727cf3a8c98e5d9c3f