Fix nullability of suspend function's Continuation parameter
Kotlin adds a synthetic `kotlin.coroutines.Continuation` parameter to
a `suspend` function. The `Continuation` type takes a single type
argument which is the return type of the source `suspend` function.
Previously, the nullability of that type argument was not handled
correctly, it was assumed to always be non-null. This change fixes
that.

Bug: 325454211
Test: ./gradlew
Change-Id: Id728df6a607b5fe7de4c761ad4461ee289307a2d