Allow merging of duplicate classes with different annotations
Instead of throwing an error when duplicate classes with different
annotations are found, support merging by adding the new annotations to
the existing class. Given that the api signature files are guaranteed to
be passed in the order from the narrowest api scope to the widest api
scope, it is safe to add a new annotation to the existing class.

Note that this change does not fully handle merging annotations with
parameters. For instance, if the public api class possess
`@SomeAnnotation(key=val1)` and the system api class possess
`@SomeAnnotation(key=val2)`, both annotations will be added to the class
instead of the system api class annotation overriding the existing
public api class annotation.

Test: ./gradlew test
Bug: 328529959
Change-Id: I605094fdbec564152af030495c62fc3ca096e6c5