Simplify handling of lambda's type arguments
Previously, the `KotlinTypeInfo.forTypeArgument(...)` had special
handling for the `KtFunctional` that was needed for handling lambda
types. It would map the type argument index back to the corresponding
component of the `KtFunctionalType`. This change adds a new
`overrideTypeArguments` to `KotlinTypeInfo` that allows the lambda to
construct the correct set of `KotlinTypeInfo` objects for the
underlying `KtFunctionalType`.

This simplifies the `KotlinTypeInfo` and makes the follow-up change
that fixes the nullability issues with the suspend lambda simpler.

Bug: 325454211
Test: ./gradlew
Change-Id: I2b4e7e92ea4e53ba2aaee5cd6c29a11bc668ff87