Fix nullability issue with abstract properties
Previously, it was using `KtExpression.getKtType()` to get the type
from a `KtPropertyAccessor` but that returned `Unit` for some cases.
It seemed to have the correct nullability but because it was not the
correct type it could not provide the correct nullability for type
arguments to generic types. Switching to use
`KtDeclaration.getReturnKtType()` fixed the issue.

Bug: 325454211
Test: ./gradlew
Change-Id: I1f4b65ca5ea2b2984789bd5422b42a111cadfa5f