Pass the PsiElement context into fromKtElement(...)
Multiple different `PsiElement` subclasses can be generated from the
same `KtElement` and they may require different types, e.g. a
`KtProperty` can generate a backing field, setter and getter in the
`PsiElement` model. The backing field type will be different to the
setter and getter types when the property is delegated but at the
moment Metalava does not handle that correctly.

The `context` parameter is currently unused but a follow-up change
will use the `context` to fix the backing field issue.

Bug: 325454211
Test: ./gradlew
Change-Id: I2cef74650de546e7837c3c96748dfeab66ac9db4