Obtain KtType for KtPropertyAccessor from its KtProperty
Previously, while attempting to get a `KtType` from a
`KtPropertyAccessor`, `KotlinTypeInfo` used `getReturnKtType()`. While
that worked when `KtPropertyAccessor` was a getter it would return
`Unit` when it was a setter. That caused issues with the nullability of
type arguments if the property was a generic type, e.g. `List`.

This fixes that by always getting the type of the `KtPropertyAccessor`
from its `KtProperty`. That works because the setter parameter and
getter return type are currently always the same type as the property.

Bug: 325454211
Test: ./gradlew
Change-Id: I78572e193a9cc2f3cc886074ca39700ff9eee7de