Test inheritMethodFromNonApiAncestor(...) with declaration nullability annotations
Adds tests for `ClassItem.inheritMethodFromNonApiAncestor(...)` with
declaration nullability annotations. This is needed to replicate a use
case in AndroidX where because of the lack of type use nullability
annotations it uses declaration nullability annotations in a way that
is not strictly correct.

In Kotlin, type parameter `T` in `Foo<T: Any?>` could be null or could
be non-null depending on the type that is passed. In `Foo<String>` then
`T` is non-null, in `Foo<String?>` then `T` is null. That means `T`
has an undefined nullability and cannot be assigned `null`. However,
`T?` is nullable and can be assigned `null`. If `T` was defined as
`T: Any` then it could never be `null` and so uses of it would always
be non-null.

In Java, without type nullability annotations, it is not possible to
define a type parameter `T` as being non-null. It would always have an
undefined nullability unless it was the type of a declaration that
was annotated with `@Nullable` in which case it would be nullable.

Strictly speaking, annotating type parameter uses with `@NonNull` would
not make any sense as its nullability is determined by the type that is
passed to it. However, as AndroidX does not currently have access to
type use nullability annotations it does sometimes use `@NonNull T` to
indicate that `T` should be non-null.

e.g. see https://cs.android.com/android/platform/superproject/+/androidx-main:frameworks/support/tvprovider/tvprovider/src/main/java/androidx/tvprovider/media/tv/BaseProgram.java;l=922;drc=187e9a2088c3f281cfb617b9bedd94b9a3546d8b

This change adds a test to ensure that behavior is preserved.

Bug: 325454211
Test: ./gradlew
Change-Id: I7ecb6e1140cc2a0ad409ecb2af72218dddb3b4c2